reput PropagationStrategy
