+++ nncf/common/quantization/quantizer_propagation/solver.py
@@ -33,6 +33,7 @@ from nncf.common.quantization.quantizer_propagation.grouping import QuantizersWa
 from nncf.common.quantization.quantizer_propagation.structs import IgnoreReason
 from nncf.common.quantization.quantizer_propagation.structs import PropagatingQuantizer
 from nncf.common.quantization.quantizer_propagation.structs import PropagationPath
+from nncf.common.quantization.quantizer_propagation.structs import PropagationStrategy
 from nncf.common.quantization.quantizer_propagation.structs import QuantizationTrait
 from nncf.common.quantization.quantizer_propagation.structs import QuantizerPropagationStateGraphNodeType
 from nncf.common.quantization.quantizer_setup import DEFAULT_QUANTIZER_CONFIG
@@ -57,25 +58,6 @@ class TransitionStatus(Enum):
     SHOULD_MERGE = 1
     SHOULD_NOT_TRANSITION = 2
     SHOULD_WAIT_FOR_MERGE = 3
-
-
-class PropagationStrategy(Enum):
-    # While propagating up through a downward-branching node:
-    # ... do not merge at all
-    DO_NOT_MERGE_BRANCH_FQS = 0
-
-    # ... only merge for exact configuration space matches
-    MERGE_IF_ALL_BRANCH_FQ_OPTIONS_SAME = 1
-
-    # ... merge common parts, and if a branch quantizer has options for
-    # narrowing (bitwidth/mode/per-channel/etc.) in addition to
-    # the common part, keep the quantizer on branch
-    MERGE_WITH_POTENTIAL_REQUANTIZATION = 2
-
-    # ... merge common config options into a single config space for the global FQ,
-    # do not merge if this is impossible for the current branching situation and given
-    # HW config file
-    MERGE_WITH_SINGLE_FQ_RESULT = 3
 
 
 class FinalizedQuantizationProposal:
